updating changes

+++ src/main/java/com/sda/course/project/restaurant/service/RoleService.java
@@ -6,14 +6,15 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class RoleService {
     @Autowired
     private RoleRepository roleRepository;
 
-    public RoleEntity getById(Integer id) {
-        return roleRepository.findById(id).orElse(null);
+    public Optional<RoleEntity> getById(Integer id) {
+        return roleRepository.findById(id);
     }
 
     public List<RoleEntity> getAllRoles() {
@@ -24,12 +25,26 @@ public class RoleService {
         return roleRepository.save(role);
     }
 
-    public void deleteRoleById(Integer id) {
-        roleRepository.deleteById(id);
+    public Optional<RoleEntity> deleteRoleById(Integer id) {
+        Optional<RoleEntity> roleToDelete = roleRepository.findById(id);
+
+        if (roleToDelete.isPresent()) {
+            roleRepository.deleteById(id);
+        }
+        return roleToDelete;
     }
 
-    public RoleEntity updateRole(RoleEntity updatedRole) {
-        return roleRepository.save(updatedRole);
+    public RoleEntity updateRole(Integer id, RoleEntity updatedRole) {
+        return roleRepository.findById(id)
+                .map(r -> {
+                    r.setName(updatedRole.getName());
+
+                    return roleRepository.save(r);
+                }).orElseGet(() -> {
+                    updatedRole.setId(id);
+
+                    return roleRepository.save(updatedRole);
+                });
     }
 
 }
